Add embed_query_prefix option for embedding models like e5
Some embedding models, like e5, require queries and passages to be
prefixed with strings like 'query: ' and 'passage: ' to obtain the best
results. This commit adds a parameter to Docs which allows settings the
query embedding search prefix, a prefix can be added to passages by
manually chunking and embedding them before handing them to paper-qa.