Make OpenAlex work (#1168)
It turns out openalex never worked if you query on DOI. OpenAlex refers to DOIs always with the URL prefix (e.g., https://doi.org/10xxx/xxxxx). So when it returned articles, they had DOIs that didn't match the requested DOIs and we rejected them as failed. This fixes and adds a low-effort unit test that just checks if openalex is working.